Refactor: rename constants directory
