+++ src/driveSemaphore.ts
@@ -3,7 +3,7 @@ import path from 'node:path';
 import async, { AsyncResultCallback } from 'async';
 import { Mutex, Semaphore } from 'async-mutex';
 
-import Defaults from './constants/defaults.js';
+import Defaults from './globals/defaults.js';
 import FsPoly from './polyfill/fsPoly.js';
 import File from './types/files/file.js';
