+++ src/igir.ts
@@ -8,8 +8,8 @@ import isAdmin from 'is-admin';
 import Logger from './console/logger.js';
 import ProgressBar, { ProgressBarSymbol } from './console/progressBar.js';
 import ProgressBarCLI from './console/progressBarCli.js';
-import Defaults from './constants/defaults.js';
-import Package from './constants/package.js';
+import Defaults from './globals/defaults.js';
+import Package from './globals/package.js';
 import CandidateArchiveFileHasher from './modules/candidateArchiveFileHasher.js';
 import CandidateCombiner from './modules/candidateCombiner.js';
 import CandidateGenerator from './modules/candidateGenerator.js';
