+++ src/modules/scanner.ts
@@ -1,7 +1,7 @@
 import ProgressBar from '../console/progressBar.js';
-import Defaults from '../constants/defaults.js';
 import DriveSemaphore from '../driveSemaphore.js';
 import ElasticSemaphore from '../elasticSemaphore.js';
+import Defaults from '../globals/defaults.js';
 import ArrayPoly from '../polyfill/arrayPoly.js';
 import fsPoly from '../polyfill/fsPoly.js';
 import File from '../types/files/file.js';
