+++ src/modules/updateChecker.ts
@@ -5,7 +5,7 @@ import semver from 'semver';
 import Logger from '../console/logger.js';
 import LogLevel from '../console/logLevel.js';
 import ProgressBarCLI from '../console/progressBarCli.js';
-import Package from '../constants/package.js';
+import Package from '../globals/package.js';
 import BufferPoly from '../polyfill/bufferPoly.js';
 
 /**
