+++ src/types/files/file.ts
@@ -7,7 +7,7 @@ import {
   Exclude, Expose, instanceToPlain, plainToClassFromExist,
 } from 'class-transformer';
 
-import Defaults from '../../constants/defaults.js';
+import Defaults from '../../globals/defaults.js';
 import FilePoly from '../../polyfill/filePoly.js';
 import fsPoly from '../../polyfill/fsPoly.js';
 import URLPoly from '../../polyfill/urlPoly.js';
