+++ test/modules/romScanner.test.ts
@@ -1,7 +1,7 @@
 import os from 'node:os';
 import path from 'node:path';
 
-import Defaults from '../../src/constants/defaults.js';
+import Defaults from '../../src/globals/defaults.js';
 import ROMScanner from '../../src/modules/romScanner.js';
 import fsPoly from '../../src/polyfill/fsPoly.js';
 import Options from '../../src/types/options.js';
