+++ test/types/patches/apsGbaPatch.test.ts
@@ -1,7 +1,7 @@
 import fs from 'node:fs';
 import path from 'node:path';
 
-import Defaults from '../../../src/constants/defaults.js';
+import Defaults from '../../../src/globals/defaults.js';
 import bufferPoly from '../../../src/polyfill/bufferPoly.js';
 import fsPoly from '../../../src/polyfill/fsPoly.js';
 import File from '../../../src/types/files/file.js';
